Revert Rename "main.controllers" to "main.main_views"
