+++ adaptive_hockey_federation/analytics/views.py
@@ -4,7 +4,7 @@ from analytics.forms import AnalyticsFilterForm
 from core.constants import GENDER_CHOICES
 from dateutil.relativedelta import relativedelta
 from django.db.models import Q
-from main.main_views.player_views import PlayersListView
+from main.controllers.player_views import PlayersListView
 from main.models import Nosology, Team
