+++ adaptive_hockey_federation/main/controllers/main_views.py
